Made good progress on lazy loading
unfortunately there are problems when you try to use it on the shop page.

The first problem was in E-commerce, where we usort() and by doing that load the units. I've fixed this by just loading them before calling sort (for now).

The other issue is documented on my notepad. HAVE FUN